+++ klient/eter.h
@@ -3,11 +3,12 @@
 
 #include<vector>
 #include"field.h"
+#include<stdint.h>
 
 class Ether
 {
 public:
-    Ether(int fields_x, int fields_y, double fieldXSize, double fieldYSize);
+    Ether(int32_t fields_x, int32_t fields_y, int32_t fieldXSize, int32_t fieldYSize);
     void start();
 private:
     std::vector<Field> fields;
